+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
     "extends": "ego",
     "rules": {
+        "import/no-default-export": ["off"],
         "no-underscore-dangle": ["off"],
         "@typescript-eslint/naming-convention": ["off"]
     }