+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 7.0.0
 
+* !!! BREAKING CHANGES !!!
+  * add `allowDefaultImports` to [IInitControllersOptions](https://egodigital.github.io/express-controllers/interfaces/_index_.iinitcontrollersoptions.html) interface, which is `(true)` by default
 * miminum requirement is [Node 12+](https://nodejs.org/dist/latest-v12.x/docs/api/) now
 * code build in [strict mode](https://www.typescriptlang.org/docs/handbook/compiler-options.html) now